FIX(client): Do not abort processing sound data when master volume is low
For many years the mix() method of AudioOutput contained an early return
when the settings master volume was close to 0.

Since the mix() method became more and more complex, this early return
has caused some logic to not work correctly in that situation.
Most notably, the user talking state icons are not being updated despite
clients talking and sending packets.

This commit removes the early return with the expectation of no side-effects
being discovered, since numerically the mix() method should handle
a volume of 0.0 gracefully anyway.

Fixes #6086