FEAT(client): Display "talking muted" state when volume is low
This commit adds a check for every user whether their individual volume
is less than a certain threshold.
The individual volume is composed of settings master volume, the local
volume adjustment, as well as positional and priority volume factors.
If so, the "talking muted" status icon is being displayed instead of
the regular talking status icon.

This should help indicate to users that their combination of settings
have made users (almost) inaudible, in effect similar to being "locally muted".