FIX(server): Make msgQueryUsers perform a permission check
The QueryUsers protocol is used to fill the ACLEditor with usernames.
It relates registered user ids with registered user names, even if
the target user is offline.

Since its inception, msgQueryUsers did not check for any permissions
when replying to user requests. So in theory, any fully unpriviledged
user could get the entire registered user list if they knew what they
were doing.

This commit adds a check so that the server only responds to msgQueryUsers
if the requesting user has the Write ACL in any channel on the server.

Note: Being able to create and then creating a temporary channel is enough
to satisfy this condition.

Fixes #3394