CHANGE(client): Only display "muted"/"deafened" message from users in own channel
Since b83316ad Mumble has been displaying a "user is now muted/deafened"
notification for users in all linked channels. While this is generally not
a bad idea, this message is not checked against a speak permissions in the
target channel.

In scenarios where a large number of users are in a certain channel
but unable to speak in a linked channel, it is easy to imagine that those
mute/deaf notifications might get overwhelming for the users in the
linked channel.

We are currently not able to find a satisfying solution to check
whether or not a certain user is able to speak in a certain channel on
the client side. This is why this commit removes the mute/deaf notification
for users in linked channels as we imagine that this use-case is
more important than the current implementation.

Closes #3026