FEAT(client,server): Make banning IP/certificate hash optional
Previously, any ban issued would include both IP and certificate hash.
This is undesirable, because IP addresses may be dynamically allocated
by the users ISP. Different users might also share the same network
connection or VPN/proxy servers. Ideally, we want to be able to choose
whether to ban IP, certificate hash or both.

This commit extends the Mumble network protocol to include flags
for banning IP and/or certificate hash. This is also applied to
the client by introducing a custom BanDialog which includes checkboxes
for both ban methods, as long as the server reports a version
which supports this feature.

In case either client or server does not support this feature, the
fallback is always how it worked before: Unconditionally banning both
IP and certificate hashes.

The BanEditor is also updated to accommodate the new logic.

Closes #4621