FIX(server): Prevent negative public server list registration timer
The public server list registration is being performed on a timer.
This timer is started with a small, random deviation.

Due to an integer casting oversight, we managed to create a
situation where sometimes that timer could be started with a negative
interval leading to warnings and, in the worst case, the server
stop trying to register with the public server list.

This commit replaces the call to ``QRandomGenerator::generate`` with
a sane ``QRandomGenerator::bounded`` to fix negative timer values.

Fixes #6956