FIX(server): Reintroduce the evaluation of bRememberChan
Before 2907d6dd39d1ea3ac364cca06af70c9e75754030, the ``ServerDB`` abstraction contained
a check for ``bRememberChan``. This check was omitted when switching to the new
``DBWrapper`` abstraction.
This lead to ``rememberchannel`` being effectively ``true`` in all cases.

This commit re-introduces the check for ``bRememberChan``, this time to
``Messages.cpp`` where it makes more sense.