FIX(client): Filter and drop excessive analog input events
Previously, we would add all input events to a fixed-size
blocking Queue on Windows.
This lead to the problem where analog input devices which
create a lot of events would fill up that blocking queue
quickly, stalling the main thread and UI.

This commit replaces the "emplace" call for the queue with
"try_emplace" dropping anything that would exceed the limit
of the queue.

Furthermore, this commit adds a throttle mechanism. If we
fail to place an event in the queue, we increase a threshold value.
The threshold value determines how many intermediate
events will be unconditionally dropped before we accept a
new event from an input device (excluding mouse or keyboard).
If the queue becomes less full, the threshold value will decrease
again - hopefully balancing the processing of the queue.

Fixes #6696

Co-Authored-By: unknown <bram@referox.com>