FIX(database): Drop conflicting indices & triggers before recreation
When migrating tables it can occur that the old tables use indices and
triggers that use the same name as the ones we would like to create for
the new tables. For DBs like SQLite where index & trigger names are
global instead of being scoped to their associated table, this will
cause a conflict.

Hence, we first check if those already exist and if they do, we drop
them from the old tables in order to be able to create the new ones.

Fixes #6935