FIX(db): Explicitly handle FK dependencies for SQLite
When dropping old tables after a migration, SQLite does not error when
attempting to drop a table that is still referenced via a foreign key
(FK) from another table. Hence, the strategy of attempting to delete
tables and seeing whether the RDBMS complains is not viable to
implicitly detect the order in which tables may be deleted.

Hence, for SQLite we have to explicitly query existing FK constraints
and then determine the deletion order based off that.

See also: https://sqlite.org/forum/info/c6bed1755422cf03