FIX(db): Work around MySQL implicit commits
As it turns out, MySQL simply performs implicit commits when executing
certain queries, particularly any kind of DDL statment. For a list of
commands see https://dev.mysql.com/doc/refman/8.4/en/implicit-commit.html

This is problematic as it completely kills our idea of having control
over what statements are in a transaction. Worse still, we have no
(real) way of getting notified of this. Hence, we have to guess at which
statements will terminate our transaction and then we have to explicitly
make our transaction-tracking system aware of this fact in order to
re-establish a transaction.