feat(proto): add a UDP variant of `Authenticate`
The current flow for connection handling of mumble client puts a bunch of
un-needed strain on the server as we have to go through the entire list of
UDP sockets that aren't yet authenticated.

This instead allows us to send a partially unencrypted packet with a `session_id`
that the calling client is trying to authenticate for, and an encrypted Audio
packet that the server will use to verify that they are actually the client.