FIX(client): Enable user to reset database path, if database is missing
When the location of the Mumble client database changed, Mumble would
display an error message informing the user that the database was missing
and suggest that it would create a new database at the given location.

However, this may not be what the user wants. More often than not
you would instead like to reset the configured location and check again
or create a new database at the default path.

This commit extends the "database not found" dialog with a cancel option.

If the user selects "Yes", a database will be created at the (old) configured
path.

If the user selects "No", the configured path will be reset and a database will
be opened or created at the (new) default path.

If the user selects "Cancel", we quit without changes.

Fixes #7020